Refactor: Rename BubbleSortPlayer to ArrayVisualizer
